requirement.txt
pandas
openpyxl
matplotlib
reportlab
python-pptx
python-docx
streamlit 
plotly
pillow
